fix(fastify,next): address inline review findings
- Revert max-lines to 300, extract exchange/template helpers
- Remove oauth_not_configured toast, use oauth_failed_google for Google callback
- Fix no-throw-literal in oauth-google
